Move Bento facade to a folder Facades
- That way, Bento will be proposed by editor if we try to add missing use
