+++ src/Facades/Bento.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Exolnet\Bento;
+namespace Exolnet\Bento\Facades;
 
 use Illuminate\Support\Facades\Facade;
 
 /**
  * @see \Exolnet\Bento\Bento
  */
-class BentoFacade extends Facade
+class Bento extends Facade
 {
     /**
      * Get the registered name of the component.
